% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cashflows.R
\name{data_hedged_payoff}
\alias{data_hedged_payoff}
\title{Create a dataset that can be used to plot hedged vs unhedged cashflows}
\usage{
data_hedged_payoff(
  payoff,
  model,
  params,
  put = TRUE,
  type = "model",
  premium = "P",
  exact_daily_premium = FALSE
)
}
\description{
Create a dataset that can be used to plot hedged vs unhedged cashflows
}
\examples{
payoff <- SampleData$payoffs
model <- SampleData$model
params <- solarOption_contracts(payoff)
df_hedged <- data_hedged_payoff(payoff, model, params)
}
