% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_cashflows.R
\name{fig_cashflows}
\alias{fig_cashflows}
\alias{fig_cashflows_hedged_year}
\alias{fig_cashflows_unhedged_year}
\alias{fig_cashflows_monthly_moments}
\alias{fig_cashflows_unhedged_vs_hedged}
\alias{fig_cashflows_unhedged_vs_hedged_old}
\alias{fig_cashflows_unhedged_vs_hedged_density_year}
\alias{fig_cashflows_unhedged_vs_hedged_density_year_month}
\alias{fig_cashflows_unhedged_vs_hedged_density_year_months}
\title{Create a dataset that can be used to plot hedged vs unhedged cashflows}
\usage{
fig_cashflows_hedged_year(df_hedged, nyear)

fig_cashflows_unhedged_year(df_hedged, nyear)

fig_cashflows_monthly_moments(df_hedged, nyear)

fig_cashflows_unhedged_vs_hedged(df_hedged, nyear)

fig_cashflows_unhedged_vs_hedged_old(df_hedged, nyear, subtitle = NULL)

fig_cashflows_unhedged_vs_hedged_density_year(
  df_hedged,
  nyear,
  subtitle = NULL
)

fig_cashflows_unhedged_vs_hedged_density_year_month(
  df_hedged,
  nyear,
  nmonth = 1
)

fig_cashflows_unhedged_vs_hedged_density_year_months(
  df_hedged,
  nyear,
  nmonth = 1
)
}
\description{
Create a dataset that can be used to plot hedged vs unhedged cashflows
}
\examples{
payoff <- SampleData$payoffs
model <- SampleData$model
params <- solarOption_contracts(payoff)
df_hedged <- data_hedged_payoff(payoff, model, params)

# Plot hedged cash flows
fig_cashflows_hedged_year(df_hedged)
fig_cashflows_hedged_year(df_hedged, 2018)
fig_cashflows_hedged_year(df_hedged, 2021)

# Plot unhedged cash flows
fig_cashflows_unhedged_year(df_hedged)
fig_cashflows_unhedged_year(df_hedged, 2018)
fig_cashflows_unhedged_year(df_hedged, 2021)

# Plot monthly mean and std. deviation of cash flows
fig_cashflows_monthly_moments(df_hedged)
fig_cashflows_monthly_moments(df_hedged, 2018)
fig_cashflows_monthly_moments(df_hedged, 2021)

# Compare unhedged and hedged cash flows
fig_cashflows_unhedged_vs_hedged(df_hedged)
fig_cashflows_unhedged_vs_hedged(df_hedged, 2018)
fig_cashflows_unhedged_vs_hedged(df_hedged, 2021)

# Compare unhedged and hedged cash flows (old version)
fig_cashflows_unhedged_vs_hedged_old(df_hedged)
fig_cashflows_unhedged_vs_hedged_old(df_hedged, 2018)
fig_cashflows_unhedged_vs_hedged_old(df_hedged, 2021)

# Compare yearly density of unhedged and hedged cash flows
fig_cashflows_unhedged_vs_hedged_density_year(df_hedged)
fig_cashflows_unhedged_vs_hedged_density_year(df_hedged, 2018)
fig_cashflows_unhedged_vs_hedged_density_year(df_hedged, 2021)

# Compare monthly density of unhedged and hedged cash flows
fig_cashflows_unhedged_vs_hedged_density_year_month(df_hedged, 2020, 4)
fig_cashflows_unhedged_vs_hedged_density_year_month(df_hedged, 2018, 1)
fig_cashflows_unhedged_vs_hedged_density_year_month(df_hedged, 2021, 2)

# Plot all monthly density of unhedged and hedged cash flows
fig_cashflows_unhedged_vs_hedged_density_year_months(df_hedged, 2018)

}
