% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_compare_mixture_parameters.R
\name{fig_compare_mixture_parameters}
\alias{fig_compare_mixture_parameters}
\title{Visually compare the Gaussian mixture parameters between two models}
\usage{
fig_compare_mixture_parameters(model_A, model_B)
}
\arguments{
\item{model_A}{first model}

\item{model_B}{second model}
}
\description{
Visually compare the Gaussian mixture parameters between two models
}
