% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_compare_sim_vs_emp.R
\name{fig_compare_sim_vs_emp}
\alias{fig_compare_sim_vs_emp}
\alias{fig_compare_sim_vs_emp_pdf}
\alias{fig_compare_sim_vs_emp_cdf}
\alias{fig_compare_sim_vs_emp_pdf_monthly}
\title{Compare simulated and empirical densities}
\usage{
fig_compare_sim_vs_emp_pdf(scenario, nmonth = 1:12, nsim = 100, bins = 20)

fig_compare_sim_vs_emp_cdf(scenario, nmonth = 1:12, nsim = 100)

fig_compare_sim_vs_emp_pdf_monthly(scenario, nsim = 100, bins = 17)
}
\description{
Compare simulated and empirical densities
}
\examples{
scenario <- SampleData$scenarios$P
# Compare Monte Carlo pdf vs empirical pdf
fig_compare_sim_vs_emp_pdf(scenario, bins = 25)

# Compare Monte Carlo pdf vs empirical pdf for every month
fig_compare_sim_vs_emp_pdf_monthly(scenario)

# Compare Monte Carlo cdf vs empirical cdf
fig_compare_sim_vs_emp_cdf(scenario)

}
