% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_ghi_year.R
\name{fig_ghi_year}
\alias{fig_ghi_year}
\title{Plot the time series of solar radiation}
\usage{
fig_ghi_year(
  model,
  nyear = 2022,
  line = TRUE,
  interval = NA,
  size = 0.3,
  alpha = 0.5,
  labels = FALSE,
  subtitle = NULL,
  seasonal_mean = FALSE,
  clearsky = TRUE,
  legend.position = c(0.98, 0.98),
  legend.justification = c("right", "top")
)
}
\arguments{
\item{model}{`solarModel` object}

\item{nyear}{reference year to plot.}

\item{line}{logical, when TRUE the GHI points will be connected with a line.}

\item{interval}{character vector, optional vector with start and end date for plotting.}

\item{size}{size of the GHI points.}

\item{alpha}{transparency for the GHI line and points.}

\item{labels}{logical, when `TRUE` will be added labels on x and y axis.}

\item{subtitle}{plot subtitle}

\item{seasonal_mean}{logical, when `TRUE` the seasonal mean will be plotted.}

\item{clearsky}{logical, when `TRUE` the seasonal clearsky will be plotted.}

\item{legend.position}{position of the legend}

\item{legend.justification}{justification of the legend.}
}
\description{
Plot the time series of solar radiation with clear sky and seasonal radiation.
}
\examples{
model <- SampleData$model
fig_ghi_year(model, nyear = 2019)
fig_ghi_year(model, line=FALSE, legend.position="none")
fig_ghi_year(model, nyear = 2019, seasonal_mean = FALSE)
fig_ghi_year(model, nyear = 2019, seasonal_mean = FALSE, clearsky = FALSE)
fig_ghi_year(model, interval = c("2020-04-01", "2020-09-22"))
fig_ghi_year(model, interval = c("2020-04-01", "2020-09-22"),
legend.position = "none", legend.justification = NULL, labels = TRUE)
}
