% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_grid_neighborhoods.R
\name{fig_grid_neighborhoods}
\alias{fig_grid_neighborhoods}
\title{Grid neighborhoods}
\usage{
fig_grid_neighborhoods(location, best_locations)
}
\arguments{
\item{location}{locations dataset}

\item{best_locations}{neighborhoods dataset.}
}
\description{
Grid neighborhoods
}
