% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_mixture_pdf.R
\name{fig_mixture_pdf}
\alias{fig_mixture_pdf}
\alias{fig_mixture_pdf_month}
\title{Plot the monthly Gaussian mixture density divided by its components}
\usage{
fig_mixture_pdf(model, bins = 30)

fig_mixture_pdf_month(model, nmonth = 1, bins = 30)
}
\arguments{
\item{model}{solarModel}

\item{bins}{number of bins}

\item{nmonth}{number of month}
}
\description{
Plot the monthly Gaussian mixture density divided by its components
}
\examples{
model <- SampleData$model
fig_mixture_pdf(model)
fig_mixture_pdf_month(model)

}
