% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_option_cum_net_payoff.R
\name{fig_option_cum_net_payoff}
\alias{fig_option_cum_net_payoff}
\title{Plot the cumulated net payoff}
\usage{
fig_option_cum_net_payoff(
  payoff,
  type = "scenarios",
  put = TRUE,
  tick = 1,
  exact_daily_premium = FALSE,
  subtitle = NULL
)
}
\description{
Plot the cumulated net payoff
}
\examples{
payoffs <- SampleData$payoffs
fig_option_cum_net_payoff(payoffs, type = "model", put = TRUE)
fig_option_cum_net_payoff(payoffs, type = "scenarios", put = TRUE)
fig_option_cum_net_payoff(payoffs, type = "model", put = FALSE)
fig_option_cum_net_payoff(payoffs, type = "scenarios", put = FALSE)
}
