% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_option_exercise_month.R
\name{fig_option_exercise_month}
\alias{fig_option_exercise_month}
\title{Figure with monthly probability of exercise under different measures}
\usage{
fig_option_exercise_month(
  payoff,
  type = "scenarios",
  put = TRUE,
  limits = c(0.2, 0.65),
  subtitle = NULL
)
}
\arguments{
\item{payoff}{object of the class `solarOptionPayoff`}

\item{type}{character, type of computation. Can be `model` or `scenarios`.}

\item{put}{logical, when `TRUE` the exercise probability refers to a PUT contract otherwise to a CALL.}

\item{limits}{numeric vector, limits for y-axis.}

\item{subtitle}{character, plot subtitle.}
}
\description{
Figure with monthly probability of exercise under different measures
}
\examples{
payoffs <- SampleData$payoffs
type = "model"
put = TRUE
limits = c(0.27, 0.58)
subtitle = NULL
fig_option_exercise_month(payoffs, type = "scenarios", put = TRUE, limits = c(0.27, 0.68), subtitle = NULL)
fig_option_exercise_month(payoffs, type = "model", put = TRUE, limits = c(0.27, 0.58), subtitle = NULL)
fig_option_exercise_month(payoffs, type = "scenarios", put = FALSE, limits = c(0.27, 0.78), subtitle = NULL)
fig_option_exercise_month(payoffs, type = "model", put = FALSE, limits = c(0.27, 0.78), subtitle = NULL)
}
