% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_option_premium_month.R
\name{fig_option_premium_month}
\alias{fig_option_premium_month}
\title{Plot the option price for every month under different measures.}
\usage{
fig_option_premium_month(
  payoffs,
  type = "scenarios",
  put = TRUE,
  tick = 1,
  subtitle = NULL
)
}
\description{
Plot the option price for every month under different measures.
}
\examples{
payoffs <- SampleData$payoffs
fig_option_premium_month(payoffs, type = "scenarios")
fig_option_premium_month(payoffs, type = "model")
fig_option_premium_month(payoffs, type = "scenarios", put = FALSE)
fig_option_premium_month(payoffs, type = "model", put = FALSE)
}
