% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_probabilistic_forecast.R
\name{fig_probabilistic_forecast}
\alias{fig_probabilistic_forecast}
\title{Plot a probabilistic forecast}
\usage{
fig_probabilistic_forecast(
  model,
  date = "2021-05-29",
  ci = 0.1,
  type = "mix",
  unconditional = FALSE
)
}
\arguments{
\item{model}{par}

\item{date}{par}

\item{ci}{par}

\item{type}{par}

\item{unconditional}{par}
}
\description{
Plot a probabilistic forecast
}
\examples{
model <- Bologna$model
fig_probabilistic_forecast(model)

}
