% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_trajectory.R
\name{fig_trajectory}
\alias{fig_trajectory}
\title{Plot a trajectory for solar radiation}
\usage{
fig_trajectory(
  model,
  scenario,
  type = c("emp", "sim"),
  VaR = 0.05,
  nyear = NULL,
  alpha = 0.05,
  nsim = 200,
  plot_nsim = 1
)
}
\description{
Plot a trajectory for solar radiation
}
\examples{
nyear = 2022
nsim = 100
pVaR = 0.05
alpha = 0.3
plot_nsim = 1
model <- SampleData$model
scenario <- SampleData$scenarios$P
fig_trajectory(model, scenario, type="emp", VaR = 0.1, nyear = 2021, nsim = 50, plot_nsim = 1, alpha = 1)
fig_trajectory(model, scenario, type="sim", VaR = 0.05, nyear = 2021, nsim = 50, plot_nsim = 1, alpha = 1)

}
