% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_theme.R
\docType{data}
\name{figure_theme}
\alias{figure_theme}
\title{Custom theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 23.
}
\usage{
figure_theme
}
\description{
Custom theme
}
\keyword{datasets}
