% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_ghi_year.R
\name{fig_ghi_year}
\alias{fig_ghi_year}
\title{Plot the time series of solar radiation}
\usage{
fig_ghi_year(
  model,
  nyear = 2022,
  interval = NA,
  size = 0.3,
  alpha = 0.5,
  labels = FALSE,
  subtitle = NULL,
  legend.position = c(0.98, 0.98),
  seasonal_mean = TRUE,
  clearsky = TRUE,
  legend.justification = c("right", "top")
)
}
\arguments{
\item{model}{`solarModel` object}

\item{subtitle}{}
}
\description{
Plot the time series of solar radiation with clear sky and seasonal radiation.
}
\examples{
library(solarr)
model <- Bologna
fig_ghi_year(model, nyear = 2019)
fig_ghi_year(model, nyear = 2019, seasonal_mean = FALSE)
fig_ghi_year(model, nyear = 2019, seasonal_mean = FALSE, clearsky = FALSE)
fig_ghi_year(model, interval = c("2020-04-01", "2020-09-22"))
fig_ghi_year(model, interval = c("2020-04-01", "2020-09-22"),
legend.position = "none", legend.justification = NULL)
}
