% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_surface.R
\name{fig_suface}
\alias{fig_suface}
\alias{fig_surface2D}
\alias{fig_surface3D}
\alias{fig_surface2D_solarModels}
\alias{fig_surface3D_solarModels}
\title{Plot surface}
\usage{
fig_surface2D(data, target = "GHI", by = 0.1)

fig_surface3D(data, target = "GHI")

fig_surface2D_solarModels(solarModels, target = "GHI", nmonth = 1, date)

fig_surface3D_solarModels(solarModels, target = "GHI", nmonth = 1, date)
}
\arguments{
\item{data}{dataset with variables `lat`, `lon` and the variable of interest specified in `target`.}

\item{target}{character, column name of the target variable to plot.}

\item{solarModels}{a list of `solarModel` objects.}
}
\description{
Plot a bi or tri dimensional grid where each square (or altitude) is filled
accordingly to the magnitude of a variable of interest.
}
\examples{
fig_surface2D(sample_data$lat_lon_GHI)
fig_surface3D(sample_data$lat_lon_GHI)
}
